+++ CHANGELOG.md
@@ -10,7 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `Family` dataset: a small ontology containing 12 axioms.
 - Unit tests up to 88% of coverage
-- DL2VecProjector includes assertion axioms with named individuals
+- `DL2VecProjector` includes assertion axioms with named individuals
+- `FastTensorDataLoader` is part of mOWL in the module `mowl.utils.data`
+- `DeepGOZero` implementation in mOWL
+- Extended `Dataset` class.
 ### Changed
 ### Deprecated
 ### Removed